refactor and fix searchable-dropdown
* Changes from prefix search to ignored case expression search which tries to
match given input in anywhere of possible items.

* Fixes #43
