+++ src/searchable-dropdown/searchable-dropdown.component.html
@@ -1,7 +1,7 @@
 <div class="btn-group" dropdown keyboardNav="true" [(isOpen)]="status.isOpen">
-  <input id="simple-btn-keyboard-nav" placeholder="{{value}}" [ngModel]="prefix" (ngModelChange)="onPrefixChange($event)" (keypress)="onKeypress($event.key)" dropdownToggle>
+  <input id="simple-btn-keyboard-nav" placeholder="{{value}}" [ngModel]="expression" (ngModelChange)="onExpressionChange($event)" (keypress)="onKeypress($event.key)" dropdownToggle>
   <ul class="dropdown-menu" dropdownMenu role="menu" aria-labelledby="simple-btn-keyboard-nav">
-    <li *ngFor="let item of items | filterByPrefix:prefix" role="menuitem">
+    <li *ngFor="let item of items | filterByExpression:expression" role="menuitem">
       <!-- href is needed for keyboard navigation -->
       <a class="dropdown-item" href="javascript:void(0)" (click)="onItemClick(item)">{{item}}</a>
     </li>
