+++ src/shared/pipes/index.ts
@@ -1,7 +1,7 @@
 import { AddAlwaysShowFieldsPipe } from './add-always-show-fields.pipe';
 import { DifferentKeysPipe } from './different-keys.pipe';
 import { ErrorsToMessagesHtmlPipe } from './errors-to-message.pipe';
-import { FilterByPrefixPipe } from './filter-by-prefix.pipe';
+import { FilterByExpressionPipe } from './filter-by-expression.pipe';
 import { FilterAndSortBySchemaPipe } from './filter-and-sort-by-schema.pipe';
 import { SelfOrEmptyPipe } from './self-or-empty.pipe';
 import { UnderscoreToSpacePipe } from './underscore-to-space.pipe';
@@ -10,7 +10,7 @@ export {
   AddAlwaysShowFieldsPipe,
   DifferentKeysPipe,
   ErrorsToMessagesHtmlPipe,
-  FilterByPrefixPipe,
+  FilterByExpressionPipe,
   FilterAndSortBySchemaPipe,
   SelfOrEmptyPipe,
   UnderscoreToSpacePipe
@@ -20,7 +20,7 @@ export const SHARED_PIPES = [
   AddAlwaysShowFieldsPipe,
   DifferentKeysPipe,
   ErrorsToMessagesHtmlPipe,
-  FilterByPrefixPipe,
+  FilterByExpressionPipe,
   FilterAndSortBySchemaPipe,
   SelfOrEmptyPipe,
   UnderscoreToSpacePipe
